Implement iomem.6  This file came from an x86 32-bit Kali box.  When we introduce it, it raised exceptions because it allocated 1-byte regions in iomem (we've never seen that before).
To resolve the issue, we significantly improved test_iomem.cpp to give us confidence that we could make a change to the algorithm.  Once we did that, we added a "Add one-byte region" feature to the iomem module.

Along the way, we added valgrind to look for (and we found) a few memory leaks.  We also significantly improved our ability to copy strings and improved iomem's string handling.

The acutal implementation of this feature involved removing 1 line of code and modifying another line of code (both ASSERTS).  The fundamental algorithm worked fine, but we didn't really *know* that until we had all of the tests built around it to prove that it was healthy.